requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3